Starting to think about what sort of backend code we'll target
My current idea is to have the compiler plugin transform the code to use
a vectorized math library so we'll be able to then just specialize this
'compile target' of sorts for each platform we want to target. This
should also still let us get compiler optimizations which I think inline
assembly usually causes problems for. We'll also be able to just use the
built in Rust explicit SIMD support when it's implemented which should
be convenient as well.