Print information about declarations as well
It looks like a syntax extension only gets explicit type declarations
and not deduced type information. Eg in the case of `let z = 0.0` we
don't know what the type of `z` is and would need the user to write
`let z: f32 = 0.0` in order for us to tell what's going on.